Limit scroll bar in sidebar to the message list of the chat view
Except when almost empty, the message list of the chat view is taller
than its available vertical space in the sidebar. Due to this a scroll
bar was shown for the whole sidebar, and everything was moved when
scrolling to see overflown messages. Now the scroll bar is shown only
for the message list, so it can be scrolled without moving the other
elements in the sidebar.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>